add delete action to table
